Numerous STAN fixes. A lot of bcmul fixes to cast to int when it returns a string, and legacy object removal from old pbx

+++ app/Models/Quote.php
@@ -198,7 +198,7 @@ class Quote extends Model
             if (!$product->frequency || !$product->payments) continue; // Only count financed
             $total += $product->frequency->splitTotal($product->qty * $product->price, $product->payments);
         }
-        return bcmul($total, 1);
+        return (int) bcmul($total, 1);
     }
 
     /**
@@ -223,7 +223,7 @@ class Quote extends Model
             if ($product->frequency && $product->payments) continue; // Don't count financed
             $total += ($product->price * $product->qty) + $product->addonTotal;
         }
-        return bcmul($total, 1);
+        return (int) bcmul($total, 1);
     }
 
     /**
